Ability to specify ruleset filename
PHPMD allows you to actually enter a filename path in the ruleset
command itself, but it requires you to enter the path.
This commit allows passing a simple filename, and if it detects a xml
filename passed without being a relative or absolute path it'll look in
the current or parent directories to find the file.

Fixes #58